refactor: made suggested changes

+++ search/saddleback_search.cpp
@@ -30,16 +30,16 @@ namespace saddleback {
  * element is present.
  * @return An std::pair with (0, 0), if the element is not present.
  */
-std::pair<uint64_t, uint64_t> saddleback(std::vector<std::vector<int32_t>> matrix,
+std::pair<uint32_t, uint32_t> saddleback(std::vector<std::vector<int32_t>> matrix,
                                int32_t element) {
-    uint64_t left_index = 0;
-    uint64_t right_index = matrix[0].size() - 1;  // Start from top right corner
+    uint32_t left_index = 0;
+    uint32_t right_index = matrix[0].size() - 1;  // Start from top right corner
     while (left_index < matrix.size()) {  // Exit once the value of indexes get out of range.
         if (element ==
             matrix[left_index]
                   [right_index]) {  // If value on this position of matrix is
                                     // equal to element, return (row, column).
-            return std::pair<uint64_t, uint64_t>(left_index+1, right_index+1);
+            return std::pair<uint32_t, uint32_t>(left_index+1, right_index+1);
         } else if (element >
                    matrix[left_index]
                          [right_index]) {  // Else if value on this position of
@@ -56,7 +56,7 @@ std::pair<uint64_t, uint64_t> saddleback(std::vector<std::vector<int32_t>> matri
             else --right_index;
         }
     }
-    return std::pair<uint64_t, uint64_t>(
+    return std::pair<uint32_t, uint32_t>(
         0, 0);  // If the program reaches here, that means one of the index
                   // went out of index, hence no element present.
 }
@@ -74,25 +74,25 @@ static void test() {
                                             {4, 40, 400, 4000, 40000},
                                             {5, 50, 500, 5000, 50000}};
 
-    std::pair<uint64_t, uint64_t> not_found = std::pair<uint64_t, uint64_t>(0, 0);
-    std::pair<uint64_t, uint64_t> test_answer;
+    std::pair<uint32_t, uint32_t> not_found = std::pair<uint32_t, uint32_t>(0, 0);
+    std::pair<uint32_t, uint32_t> test_answer;
     // Test 1
-    std::pair<uint64_t, uint64_t> answer1 = search::saddleback::saddleback(matrix, 123);
+    std::pair<uint32_t, uint32_t> answer1 = search::saddleback::saddleback(matrix, 123);
     assert(not_found == answer1);
     // Test 2
     answer1 = search::saddleback::saddleback(matrix, 0);
     assert(not_found == answer1);
     // Test 3
     answer1 = search::saddleback::saddleback(matrix, 1);
-    test_answer = std::pair<uint64_t, uint64_t>(1, 1);
+    test_answer = std::pair<uint32_t, uint32_t>(1, 1);
     assert(test_answer == answer1);
     // Test 4
     answer1 = search::saddleback::saddleback(matrix, 50000);
-    test_answer = std::pair<uint64_t, uint64_t>(5, 5);
+    test_answer = std::pair<uint32_t, uint32_t>(5, 5);
     assert(test_answer == answer1);
     // Test 5
     answer1 = search::saddleback::saddleback(matrix, 300);
-    test_answer = std::pair<uint64_t, uint64_t>(3, 3);
+    test_answer = std::pair<uint32_t, uint32_t>(3, 3);
     assert(test_answer == answer1);
 }
